Move preset management into a separate file
